refactor(register): B2B-4467 extract BC-to-B2B flow into Register component
Made-with: Cursor
